Fill out `tower-http-service`
