+++ tower-http-service/src/lib.rs
@@ -37,7 +37,6 @@ pub trait HttpService<RequestBody>: Sealed<RequestBody> {
     /// Process the request and return the response asynchronously.
     fn call(&mut self, request: Request<RequestBody>) -> Self::Future;
 
-    /*
     /// Wrap the HttpService so that it implements tower_service::Service
     /// directly.
     ///
@@ -45,40 +44,20 @@ pub trait HttpService<RequestBody>: Sealed<RequestBody> {
     /// `HttpService` instance needs to be used where a `T: Service` is
     /// required, it must be wrapped with a type that provides that
     /// implementation. `LiftService` does this.
-    fn lift(self) -> LiftService<Self> where Self: Sized {
-        LiftService { inner: self }
+    fn into_service(self) -> util::IntoService<Self> where Self: Sized {
+        util::IntoService::new(self)
     }
 
     /// Same as `lift` but operates on an HttpService reference.
-    fn lift_ref(&mut self) -> LiftServiceRef<Self> where Self: Sized {
-        LiftServiceRef { inner: self }
+    fn as_service(&mut self) -> util::AsService<Self> where Self: Sized {
+        util::AsService::new(self)
     }
-    */
 }
-
-/*
-/// Wraps an `HttpService` instance, implementing `tower_service::Service`.
-///
-/// See [`lift`] function documentation for more details.
-///
-/// [`lift`]: #
-pub struct LiftService<T> {
-    inner: T,
-}
-
-/// Wraps an `HttpService` reference, implementing `tower_service::Service`.
-///
-/// See [`lift`] function documentation for more details.
-///
-/// [`lift`]: #
-pub struct LiftServiceRef<'a, T: 'a> {
-    inner: &'a mut T,
-}
-*/
 
 impl<T, B1, B2> HttpService<B1> for T
 where
     T: Service<Request<B1>, Response = Response<B2>>,
+    B2: Body,
 {
     type ResponseBody = B2;
     type Error = T::Error;
@@ -96,40 +75,5 @@ where
 impl<T, B1, B2> Sealed<B1> for T
 where
     T: Service<Request<B1>, Response = Response<B2>>,
+    B2: Body,
 {}
-
-/*
-impl<T, ReqBody> Service<Request<ReqBody>> for LiftService<T>
-where
-    T: HttpService<ReqBody>
-{
-    type Response = Response<T::ResponseBody>;
-    type Error = T::Error;
-    type Future = T::Future;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, request: Request<ReqBody>) -> Self::Future {
-        self.inner.call(request)
-    }
-}
-
-impl<'a, T, ReqBody> Service<Request<ReqBody>> for LiftServiceRef<'a, T>
-where
-    T: HttpService<ReqBody>
-{
-    type Response = Response<T::ResponseBody>;
-    type Error = T::Error;
-    type Future = T::Future;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, request: Request<ReqBody>) -> Self::Future {
-        self.inner.call(request)
-    }
-}
-*/
