--- tower-http/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "tower-http"
-version = "0.1.0"
-authors = ["Carl Lerche <me@carllerche.com>"]
-
-[dependencies]
-# futures = "0.1"
-# http = "0.1"
-# tower-add-origin = { path = "tower-add-origin" }
-# tower-service = { git = "https://github.com/tower-rs/tower" }
